+++ RELEASE-NOTES.md
@@ -3,6 +3,12 @@ different releases and which versions of PHP and MediaWiki they support, see the
 [platform compatibility tables](INSTALL.md#platform-compatibility-and-release-status).
 
 
+## Maps 7.12.2
+
+Released on December 9th, 2019.
+
+* Invalid KML file names are no longer passed to Google Maps
+
 ## Maps 7.12.1
 
 Released on December 9th, 2019.